Moves routes to new folder Edits api routes to /api/
