+++ routes/htmlRoutes.js
@@ -1,0 +1,12 @@
+var db = require("../models");
+
+module.exports = (app) => {
+
+  app.get("/", function(req, res) {
+    res.render("index");
+  });
+
+  
+  
+
+}
